requirements.txt
# Core Frameworks
fastapi>=0.110.0
uvicorn>=0.29.0
pydantic>=2.6.0
python-dotenv>=1.0.0

# AI & Orchestration
langchain>=0.1.0
langchain-core>=0.1.0
langgraph>=0.0.20
langchain-openai
langchain-community

# Inference & Models
vllm>=0.4.0
transformers>=4.38.0
torch>=2.2.0
modelscope>=1.13.0
peft>=0.10.0
trl>=0.8.0
datasets>=2.18.0

# Vector DB & RAG
pymilvus>=2.3.0
milvus-model>=0.1.0
sentence-transformers

# Data Processing
magic-pdf>=0.6.0  # MinerU
numpy
pandas
tqdm

# Crawling
trafilatura>=1.6.0

# Tools & Utilities
trafilatura
requests
aiohttp
curl_cffi